feat: standardize config and tooling
