+++ client/src/api/profile.ts
@@ -1,4 +1,4 @@
-import { api } from '@/config/api';
+import { api } from '@/lib/http';
 
 export interface ProductData {
   _id?: string;
