+++ client/src/components/ui/OrderModal/OrderModal.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react';
 import { useSelector } from 'react-redux';
 import type { RootState } from '../../../store';
-import { api } from '@/config/api';
+import { api } from '@/lib/http';
 import ModalSheet from '../../base/ModalSheet';
 import type { Product } from '../ProductCard';
 import showToast from '../Toast';
