+++ client/src/pages/OrderNow/OrderNow.tsx
@@ -3,7 +3,7 @@ import { useDispatch, useSelector } from 'react-redux';
 import { motion } from 'framer-motion';
 import { FaMicrophone } from 'react-icons/fa';
 import { fetchShops } from '@/store/shops';
-import { api } from '@/config/api';
+import { api } from '@/lib/http';
 import type { RootState } from '../../store';
 import ModalSheet from '../../components/base/ModalSheet';
 import Loader from '../../components/Loader';
