+++ server/.env.example
@@ -1,4 +1,3 @@
-FIREBASE_PROJECT_ID=mana-city-98fa0
-FIREBASE_CLIENT_EMAIL=<service-account@mana-city-98fa0.iam.gserviceaccount.com>
-FIREBASE_PRIVATE_KEY="-----BEGIN PRIVATE KEY-----\n...\n-----END PRIVATE KEY-----\n"
-JWT_SECRET=change_me
+MONGO_URI=
+JWT_SECRET=
+CORS_ORIGIN=http://localhost:5173,https://manacity4-0-1.onrender.com
