+++ server/eslint.config.js
@@ -1,0 +1,10 @@
+module.exports = [
+  {
+    ignores: ['dist', 'node_modules'],
+  },
+  {
+    files: ['**/*.js'],
+    languageOptions: { ecmaVersion: 2020 },
+    rules: {},
+  },
+];
