+++ server/package.json
@@ -6,7 +6,8 @@
     "test": "jest",
     "start": "node server.js",
     "dev": "nodemon server.js",
-    "seed": "ts-node scripts/seed/dev.ts"
+    "seed": "ts-node scripts/seed/dev.ts",
+    "lint": "eslint ."
   },
   "keywords": [],
   "author": "",
